requirements.txt
streamlit==1.28.1
textblob==0.17.1
pyspellchecker==0.7.2
language-tool-python==2.7.1
pandas==2.1.3